Remove PasswordHistory
This is a feature that has been deprecated and can be safely removed.

DEPR-7
