+++ common/djangoapps/student/tests/test_events.py
@@ -130,7 +130,7 @@ class TestUserEvents(UserSettingsEventTestMixin, TestCase):
         """
         Verify that we don't emit events for related fields.
         """
-        self.user.passwordhistory_set.create(password='new_password')
+        self.user.loginfailures_set.create()
         self.user.save()
         self.assert_no_events_were_emitted()
